+++ Tests/AdvancedOperationTests/InjectionTests.swift
@@ -1,4 +1,4 @@
-//
+// 
 // AdvancedOperation
 //
 // Copyright © 2016-2019 Tinrobots.
@@ -24,7 +24,7 @@
 import XCTest
 @testable import AdvancedOperation
 
-class InjectionTests: XCTestCase {
+final class InjectionTests2: XCTestCase {
   func testInputAndOutput() {
     let operation1 = IntToStringOperation()
     operation1.input = 10
@@ -37,38 +37,38 @@ class InjectionTests: XCTestCase {
     XCTAssertEqual(operation2.output, 10)
   }
 
-  func testInjectionUsingClassMethodAndWaitUntilFinished() {
-    let operation1 = IntToStringOperation()
-    let operation2 = StringToIntOperation()
-    operation1.input = 10
-    let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
-    let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
-
-    XCTAssertEqual(operation2.output, 10)
-  }
-
-  func testInjectionMixedWithOtherOperation() {
-    let operation1 = IntToStringOperation()
-    let operation2 = StringToIntOperation()
-    let operation3 = BlockOperation()
-    operation3.addDependency(operation2)
-    operation1.input = 10
-    let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
-    let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
-    queue.addOperations([operation3], waitUntilFinished: false)
-
-    XCTAssertEqual(operation2.output, 10)
-  }
+//  func testInjectionUsingClassMethodAndWaitUntilFinished() {
+//    let operation1 = IntToStringOperation()
+//    let operation2 = StringToIntOperation()
+//    operation1.input = 10
+//    let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
+//    let queue = AdvancedOperationQueue()
+//    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
+//
+//    XCTAssertEqual(operation2.output, 10)
+//  }
+//
+//  func testInjectionMixedWithOtherOperation() {
+//    let operation1 = IntToStringOperation()
+//    let operation2 = StringToIntOperation()
+//    let operation3 = BlockOperation()
+//    operation3.addDependency(operation2)
+//    operation1.input = 10
+//    let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
+//    let queue = AdvancedOperationQueue()
+//    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
+//    queue.addOperations([operation3], waitUntilFinished: false)
+//
+//    XCTAssertEqual(operation2.output, 10)
+//  }
 
   func testInjectionUsingInstanceMethodAndWaitUntilFinished() {
     let operation1 = IntToStringOperation()
     let operation2 = StringToIntOperation()
     operation1.input = 10
-    let adapterOperation = operation1.injectOutput(into: operation2)
+    operation1.injectOutput(into: operation2)
     let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
+    queue.addOperations([operation1, operation2], waitUntilFinished: true)
     XCTAssertEqual(operation2.output, 10)
   }
 
@@ -76,7 +76,7 @@ class InjectionTests: XCTestCase {
     let operation1 = IntToStringOperation()
     let operation2 = IntToStringOperation()
     operation1.input = 10
-    let adapterOperation = operation1.injectOutput(into: operation2) { value -> Int? in
+    operation1.injectOutput(into: operation2) { value -> Int? in
       if let value = value {
         return Int(value)
       } else {
@@ -84,7 +84,7 @@ class InjectionTests: XCTestCase {
       }
     }
     let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
+    queue.addOperations([operation1, operation2], waitUntilFinished: true)
 
     XCTAssertEqual(operation2.output, "10")
   }
@@ -93,7 +93,7 @@ class InjectionTests: XCTestCase {
     let operation1 = IntToStringOperation()
     let operation2 = IntToStringOperation()
     operation1.input = 404
-    let adapterOperation = operation1.injectOutput(into: operation2) { value -> Int? in
+    operation1.injectOutput(into: operation2) { value -> Int? in
       if let value = value {
         return Int(value)
       } else {
@@ -101,7 +101,7 @@ class InjectionTests: XCTestCase {
       }
     }
     let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
+    queue.addOperations([operation1, operation2], waitUntilFinished: true)
 
     XCTAssertNil(operation2.output)
   }
@@ -116,9 +116,9 @@ class InjectionTests: XCTestCase {
     operation3.addDependency(operation2)
 
     operation1.input = 10
-    let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
+    operation1.injectOutput(into: operation2)
     let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation, operation3], waitUntilFinished: false)
+    queue.addOperations([operation1, operation2, operation3], waitUntilFinished: false)
 
     waitForExpectations(timeout: 3)
     XCTAssertEqual(operation2.input, "10")
@@ -133,9 +133,9 @@ class InjectionTests: XCTestCase {
       expectation1.fulfill()
     }
     operation1.input = 2000 // values greater than 1000 will produce a nil output when executing 😎
-    let adapterOperation = operation1.injectOutput(into: operation2)
+    operation1.injectOutput(into: operation2)
     let queue = AdvancedOperationQueue()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: false)
+    queue.addOperations([operation1, operation2], waitUntilFinished: false)
 
     waitForExpectations(timeout: 3)
     XCTAssertNil(operation2.input)
@@ -146,37 +146,37 @@ class InjectionTests: XCTestCase {
     let operation1 = IntToStringOperation() // no input -> fails
     let operation2 = StringToIntOperation()
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(AdvancedOperation.isFinished), object: operation2, expectedValue: true)
-    let adapterOperation = operation1.injectOutput(into: operation2)
+    operation1.injectOutput(into: operation2)
     let queue = AdvancedOperationQueue()
 
     operation1.cancel()
-    queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: false)
+    queue.addOperations([operation1, operation2], waitUntilFinished: false)
 
     wait(for: [expectation1], timeout: 10)
   }
 
-//  func testInvestigationWithStandardDependency() {
-//    let operation1 = BlockOperation {}
-//    let operation3 = BlockOperation {}
-//    let operation2 = BlockOperation { [unowned operation1, unowned operation3] in }
-//    let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: operation3, expectedValue: true)
-//
-//    operation3.addDependency(operation2)
-//    operation2.addDependency(operation1)
-//    operation1.cancel()
-//
-//    let queue = OperationQueue()
-//
-//    /// ✅works
-//    queue.addOperations([operation1, operation3, operation2], waitUntilFinished: false)
-//
-//    // ⚠️ sporadic failures (i.e. 5 failures on 10_000 executions.)
-//    // when referencing operation1 and operation3 inside operation2 (with weak or unowned)
-//
-//    // queue.addOperation(operation1)
-//    // queue.addOperation(operation3)
-//    // queue.addOperation(operation2)
-//
-//    wait(for: [expectation1], timeout: 20)
-//  }
+  //  func testInvestigationWithStandardDependency() {
+  //    let operation1 = BlockOperation {}
+  //    let operation3 = BlockOperation {}
+  //    let operation2 = BlockOperation { [unowned operation1, unowned operation3] in }
+  //    let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: operation3, expectedValue: true)
+  //
+  //    operation3.addDependency(operation2)
+  //    operation2.addDependency(operation1)
+  //    operation1.cancel()
+  //
+  //    let queue = OperationQueue()
+  //
+  //    /// ✅works
+  //    queue.addOperations([operation1, operation3, operation2], waitUntilFinished: false)
+  //
+  //    // ⚠️ sporadic failures (i.e. 5 failures on 10_000 executions.)
+  //    // when referencing operation1 and operation3 inside operation2 (with weak or unowned)
+  //
+  //    // queue.addOperation(operation1)
+  //    // queue.addOperation(operation3)
+  //    // queue.addOperation(operation2)
+  //
+  //    wait(for: [expectation1], timeout: 20)
+  //  }
 }
